text_analyzer/requirements.txt
asgiref==3.8.1
Django==5.1.7
sqlparse==0.5.0
tzdata==2025.2
psycopg2-binary==2.9.9
drf-yasg
djangorestframework
python-docx
PyPDF2
scikit-learn


